ml/requirements.txt
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.10.0
tensorflow==2.12.0
Pillow==10.1.0
numpy==1.24.3
functions-framework==3.4.0
flask==3.0.0
kagglehub>=0.2.0
